azure_function/requirements.txt
azure-functions
pymongo[srv]
requests
